dataset/pyproject.toml
[tool.poetry]
name = "dataset"
version = "0.1.0"
description = "Generating the movieclip dataset"
authors = ["Keane Moraes <lordvader3002@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
argparse = "^1.4.0"
python-dotenv = "^1.0.0"
tqdm = "^4.66.1"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
ruff = "^0.1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
